feat(openfoam): replace the project-coded tap factory with a neutral one
Removes the project-coded default-tap helper from the package surface and
replaces it with rectangular_tank_wall_taps, which derives a deterministic rake
of side-wall point taps entirely from caller-supplied dimensions and encodes no
particular tank. Names are neutral and sequential in emission order.

The removal is an intentional public break rather than a rename. No alias, no
deprecation shim and no migration mapping is shipped, because any of those would
preserve the disclosure this issue exists to remove. The factory had no caller
anywhere in the repository outside its own module, the package init and its own
tests, so the break costs no in-repo adaptation.

Also splits the 629-line tap module by responsibility, which the size rule
required once the file was being touched: the data model moves to
pressure_tap_models, the post-processing to pressure_tap_analysis, and rendering
plus the new factory stay behind. Both new modules are re-exported from the
original module path so existing import sites are unaffected.

Package exports and the existing tap tests move to the neutral factory. Full
openfoam suite shows no new failures against the recorded baseline.

Refs #1574
