+++ src/digitalmodel/solvers/openfoam/__init__.py
@@ -52,10 +52,10 @@ from .motion import (
 from .pressure_taps import (
     PressureTap,
     PressureTapStatistics,
-    b1546_default_taps,
     compute_tap_statistics,
     point_tap_names,
     read_tap_statistics,
+    rectangular_tank_wall_taps,
     render_patch_probes_entry,
     render_pressure_tap_functions,
     render_probes_entry,
@@ -141,7 +141,7 @@ __all__ = [
     "prismatic_tank_natural_frequency",
     "PressureTap",
     "PressureTapStatistics",
-    "b1546_default_taps",
+    "rectangular_tank_wall_taps",
     "compute_tap_statistics",
     "point_tap_names",
     "read_tap_statistics",
